rtc/starfruit/requirements.txt
tornado
futures
riak
motor
schematics==2.0.0a1
psycopg2
queries
numpy
scipy
pyzmq
Jinja2
pandas
arrow
pycurl
ujson
pylibmc
protobuf
msgpack-python